Cleanup xcustom.h, add riscv_test_rocc.h
This adds in a raw XCustom macro to xcustom.h. This is necessary if
you don't have a C/C++ environment and have to live with low-level
assembly (e.g., writing smoke tests). This changes some of the names
for various macros to always use the XCUSTOM convention as opposed to
CUSTOMX.

This also adds a header for handling RoCC tests, riscv_test_rocc.h,
which will set the XS bit so that we don't get an illegal instruction
exception.